requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
tensorflow
paho-mqtt
gunicorn
